Refactoring - got GUI and plotter partially separated
