+++ src/main/kotlin/com/nurflugel/sisyphus/gui/GuiPreviewer.kt
@@ -1,5 +1,8 @@
 package com.nurflugel.sisyphus.gui
 
+import com.nurflugel.sisyphus.gui.GuiUtils.Companion.aliasedRenderingHints
+import com.nurflugel.sisyphus.gui.GuiUtils.Companion.isUsable
+import com.nurflugel.sisyphus.gui.GuiUtils.Companion.maxDeltaTheta
 import org.apache.commons.io.FileUtils
 import java.awt.*
 import java.awt.Color.BLACK
@@ -31,26 +34,10 @@ private const val HEIGHT = 1440
 private const val SHUT_DOWN_WITH_KEY_PRESS = false
 
 class GuiPreviewer {
-
     private val frame = JFrame()
     private var guiPanel = JPanel()
 
     companion object {
-        const val imagesDir = "images4"
-        const val tracksDir = "tracks4"
-        const val maxDeltaTheta = 1.0 / 180.0 * PI // one degree max theta
-        val aliasedRenderingHints = run {
-            val hints = RenderingHints(KEY_ANTIALIASING, VALUE_ANTIALIAS_ON)
-            hints[KEY_RENDERING] = VALUE_RENDER_QUALITY;
-            hints[KEY_ANTIALIASING] = VALUE_ANTIALIAS_OFF;
-            hints;
-        }
-        val nonAliasedRenderingHints = run {
-            val hints = RenderingHints(KEY_ANTIALIASING, VALUE_TEXT_ANTIALIAS_OFF)
-            hints[KEY_RENDERING] = VALUE_RENDER_QUALITY
-            hints
-        }
-
         // run this to display an existing .thr track
         @JvmStatic
         fun main(args: Array<String>) {
@@ -80,9 +67,6 @@ class GuiPreviewer {
         frame.pack()
         frame.preferredSize = Dimension(WIDTH, HEIGHT) // todo need to add 45
         frame.minimumSize = Dimension(WIDTH, HEIGHT) // todo need to add 45
-        //        frame.preferredSize = Dimension(WIDTH, HEIGHT + 45) // todo need to add 45
-        //        frame.size = Dimension(WIDTH, HEIGHT + 45)
-        //        frame.size = Toolkit.getDefaultToolkit().screenSize;
         frame.extendedState = JFrame.MAXIMIZED_BOTH;
 
         println("       width from frame = ${frame.size.width}")
@@ -128,7 +112,7 @@ class GuiPreviewer {
 
         var previousPoint: Pair<Double, Double>? = null
 
-        val graphicsPairs = createGraphicsCoordinates(lines, scaleFactorY, offsetX, offsetY)
+        val graphicsPairs = GuiUtils.createGraphicsCoordinates(lines, scaleFactorY, offsetX, offsetY)
 
         val bImg = BufferedImage(guiPanel.width, guiPanel.height, TYPE_INT_RGB)
         val cg = bImg.createGraphics()
@@ -160,112 +144,18 @@ class GuiPreviewer {
 
         var previousPoint: Pair<Double, Double>? = null
 
-        val graphicsPairs = createGraphicsCoordinates(lines, scaleFactorY, offsetX, offsetY)
+        val graphicsPairs = GuiUtils.createGraphicsCoordinates(lines, scaleFactorY, offsetX, offsetY)
 
         val bImg = BufferedImage(guiPanel.width, guiPanel.height, TYPE_INT_RGB)
         val cg = bImg.createGraphics()
         guiPanel.paintAll(cg)
 
         for (currentPoint in graphicsPairs) {
-            drawImageToSave(previousPoint, currentPoint, cg, imageFileName, showNames)
+            //            drawImageToSave(previousPoint, currentPoint, cg, imageFileName, showNames)
             previousPoint = currentPoint
         }
-        val imageFile = File("./$imagesDir/${imageFileName}")
-        if (ImageIO.write(bImg, "png", imageFile)) {
-            println("$imageFile -- saved")
-        }
     }
 
-    /** This function takes the raw rho/theta pairs, cleans them up, removes
-     * any comments, then "expands" to handle converting lines between two into a series of rho/theta arcs which approximates that line.
-     *
-     * Finally, converts them into graphics coordinates suitable for plotting/printing.
-     */
-    private fun createGraphicsCoordinates(
-            lines: MutableList<String>,
-            scaleFactorY: Int, // how much to scale by.  Assumption is that scaleFactorY==scaleFactorX (i.e., equal scaling)
-            offsetX: Int,
-            offsetY: Int,
-                                         ): List<Pair<Double, Double>> {
-        val polarPairs: List<Pair<Double, Double>> = lines
-            .asSequence()
-            .filter { it.isNotBlank() }
-            .map { it.trim() }
-            .filter { ! it.startsWith("//") }
-            .filter { ! it.startsWith("#") }
-            .map { convertLineToPair(it) }
-            .filter { it != null }
-            .map { it !! }
-            .toList()
-
-        val expandedPolarPairs = handleDeltaTheta(polarPairs)
-
-        val pairs = expandedPolarPairs
-            // here is where we need to take the pairs of pairs, and deal with delta thetas
-            .map { convertToXy(it) }
-            .map { convertToScreenCoords(it, scaleFactorY, offsetX, offsetY) }
-            .filter { isUsable(it) }
-            .toList()
-        return pairs
-    }
-
-    /**
-     * for every pair and the next pair, see if the delta theta is large enough (> 1 degree or so)
-     * to need to subdivide that pair of pairs into a list of pairs.
-     *
-     * This "expands" the points, converting lines between two points into a series of rho/theta arcs which approximates that line.
-     */
-    private fun handleDeltaTheta(polarPairs: List<Pair<Double, Double>>): List<Pair<Double, Double>> {
-        val expandedPairs: MutableList<Pair<Double, Double>> = mutableListOf()
-
-        (0 until polarPairs.size - 1).forEach {
-            val here = polarPairs[it]
-            val there = polarPairs[it + 1]
-            val deltaTheta = there.first - here.first
-            val absoluteDeltaTheta = Math.abs(deltaTheta)
-            if (absoluteDeltaTheta > maxDeltaTheta) { // we need to transition theta and rho evenly between here and there
-                val expandedList: MutableList<Pair<Double, Double>> = mutableListOf()
-                // find the closest number of iterations so each delta theta approximates maxDeltaTheta
-                val sss = absoluteDeltaTheta / maxDeltaTheta
-                val toInt = sss.toInt()
-                val numberOfSplits: Int = toInt + 1
-                val newDeltaTheta = deltaTheta / numberOfSplits
-                val newDeltaRho = (there.second - here.second) / numberOfSplits
-                (0 until numberOfSplits).forEach { index ->
-                    val subTheta = here.first + (index * newDeltaTheta)
-                    val subRho = here.second + (index * newDeltaRho)
-                    expandedList.add(Pair(subTheta, subRho))
-                }
-                expandedPairs.addAll(expandedList)
-            } else { // no need to expand
-                expandedPairs.add(here)
-            }
-        }
-        expandedPairs.add(polarPairs.last())
-        return expandedPairs
-    }
-
-    /** Convert the theta/rho coordinates to XY coordinates */
-    private fun convertToXy(line: Pair<Double, Double>): Pair<Double, Double> {
-        val theta = line.first
-        val rho = line.second
-        val x = rho * cos(theta)
-        val y = rho * sin(theta)
-        return Pair(x, y)
-    }
-
-    /** Convert the XY coordinates to screen coordinates - deal with scaling and offsets*/
-    private fun convertToScreenCoords(xy: Pair<Double, Double>, scaleFactor: Int, offsetX: Int, offsetY: Int): Pair<Double, Double> {
-        return Pair(xy.first * scaleFactor + offsetX, xy.second * scaleFactor + offsetY)
-    }
-
-    private fun convertLineToPair(line: String): Pair<Double, Double>? {
-        val tokens = line.split(" ", "\t")
-        val theta = tokens[0].toDoubleOrNull()
-        val rho = tokens.last().toDoubleOrNull()
-        if (rho != null && theta != null) return Pair(theta, rho)
-        return null
-    }
 
     /** Draw a line in the context between the two points */
     private fun drawUiPreview(
@@ -292,33 +182,24 @@ class GuiPreviewer {
 }
 
 
-private fun drawImageToSave(
-        possiblePreviousPoint: Pair<Double, Double>?,
-        currentPoint: Pair<Double, Double>,
-        graphics: Graphics2D,
-        fileName: String,
-        showName: Boolean,
-                           ) {
-    val previousPoint = when {
-        isUsable(possiblePreviousPoint) -> possiblePreviousPoint !!
-        else                            -> currentPoint
-    }
-    val line = Line2D.Double(previousPoint.first, previousPoint.second, currentPoint.first, currentPoint.second)
-    graphics.color = GRAY
-    graphics.draw(line)
-    if (showName) {
-        //        graphics.setRenderingHints(nonAliasedRenderingHints)
-        graphics.drawString(fileName, 20, 30);
-        //        graphics.setRenderingHints(aliasedRenderingHints)
-    }
-}
+//private fun drawImageToSave(
+//        possiblePreviousPoint: Pair<Double, Double>?,
+//        currentPoint: Pair<Double, Double>,
+//        graphics: Graphics2D,
+//        fileName: String,
+//        showName: Boolean,
+//                           ) {
+//    val previousPoint = when {
+//        isUsable(possiblePreviousPoint) -> possiblePreviousPoint !!
+//        else                                     -> currentPoint
+//    }
+//    val line = Line2D.Double(previousPoint.first, previousPoint.second, currentPoint.first, currentPoint.second)
+//    graphics.color = GRAY
+//    graphics.draw(line)
+//    if (showName) {
+//        //        graphics.setRenderingHints(nonAliasedRenderingHints)
+//        graphics.drawString(fileName, 20, 30);
+//        //        graphics.setRenderingHints(aliasedRenderingHints)
+//    }
+//}
 
-/** Is this a usable point?  */
-private fun isUsable(pair: Pair<Double, Double>?): Boolean {
-
-    return when {
-        pair == null                              -> false
-        pair.first.isNaN() || pair.second.isNaN() -> false
-        else                                      -> true
-    }
-}
